Create project for bundling code analyzers
Create a project that produces a NuGet metapackage that bundles commonly
used analyzers for personal projects. Enforce coding standards in
consuming projects by including
- a props file that enables nullable reference types, forces the
build process to treat warnings as errors and sets the maximum analysis
level,
- a stylecop configuration file that sets the company name,
- a globalconfig file that supresses documentation-related warnings.

Add a github workflow that pushes a new version of the metapackage to
GitHub Packages on every commit.